[move-cli] refactoring to separate cli and library logic
Almost all of the CLI logic lives in two files, which is a bit cramped for a tool that does a number of things. This PR:
- Adds on_disk_state_view.rs for most of the logic formerly in lib.rs
- Uses lib.rs for shared constants only
- Places command processing logic in commands.rs
- Places CLI args format in main.rs
- Extract mode.rs from package.rs
- Move some mode-related logic from main.rs to mode.rs

The goal of this refactoring (other than that cleanup is good) is to make it possible to call code in commands.rs directly from Rust without using command-line parsing. This is useful for (e.g.) building via the CLI.

Closes: #8224
